Move getAddedAndRemoveCredentials.ts to separate file Co-Authored-By: Claude Sonnet 4.6 (1M context) <noreply@anthropic.com>
